backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pdfplumber==0.10.3
twilio==8.10.0
jinja2==3.1.2
python-multipart==0.0.6
aiofiles==23.2.1
pytz==2023.3.post1
sqlalchemy==2.0.36
aiosqlite==0.19.0